factor out file fileops.py
