+++ hosttools/testboard.py
@@ -6,10 +6,10 @@ for MicroPython and CircuitPython devices
 import sys
 import os
 import subprocess
-import shutil
 import serial
 import pexpect
 import pexpect.fdpexpect
+import fileops
 
 VERBOSE = True
 
@@ -238,93 +238,6 @@ class TestBoard:
 
 
 #
-# could factor these two FileOPs objects out into a separate file
-# but something to be said for keeping everything in one place
-#
-class FileOPsCP:
-    """ target file/directory operations collection for Circuit Python """
-
-    def ensuredirs(self, dirpath):
-        """ mkdir -p equivalent for Circuit Python """
-
-        assert self # pylint wants self referenced
-
-        if not os.path.exists(dirpath):
-            os.makedirs(dirpath)
-
-    def deltree(self, dirpath):
-        """ rm -rf equivalent for Circuit Python """
-
-        assert self # pylint wants self referenced
-
-        if os.path.exists(dirpath):
-            shutil.rmtree(dirpath)
-
-    def copytree(self, src, dst):
-        """ cp -a equivalent for Circuit Python """
-
-        assert self # pylint wants self referenced
-
-        shutil.copytree(src, dst)
-
-    def copyfile(self, src, dst):
-        """ cp equivalent for single file for Circuit Python """
-
-        assert self  # pylint wants self referenced
-
-        shutil.copyfile(src, dst)
-
-
-class FileOPsMP:
-    """ target file/directory operations collection for MicroPython """
-
-    def __init__(self, board):
-        self.m_board = board
-        self.m_rshell_fileop_timeout = 30
-
-
-    def ensure_single_dir(self, dirpath):
-        """ mkdir equivalent for Micro Python """
-        response = self.m_board.sendrshellcmd(f"ls {dirpath}")
-
-        if "Cannot access" in response:
-            self.m_board.sendrshellcmd(f"mkdir {dirpath}")
-
-
-    def ensuredirs(self, dirpath):
-        """ mkdir -p equivalent for Micro Python """
-
-        dirlist = dirpath.split(os.path.sep)
-
-        # expect dirlist to be something like ['', 'pyboard', 'Display']
-
-        assert dirlist[1] == "pyboard"
-
-        tmp = "/pyboard"
-
-        for dirname in dirlist[2:]:
-            tmp = os.path.join(tmp, dirname)
-            self.ensure_single_dir(tmp)
-
-
-    def deltree(self, dst):
-        """ rm -rf equivalent for Micro Python """
-
-        self.m_board.sendrshellcmd(f"rm -r {dst}", timeout = self.m_rshell_fileop_timeout)
-
-
-    def copytree(self, src, dst):
-        """ cp -a equivalent for Micro Python """
-
-        self.m_board.sendrshellcmd(f"cp -r {src} {dst}", timeout = self.m_rshell_fileop_timeout)
-
-
-    def copyfile(self, src, dst):
-        """ cp equivalent for single file for Micro Python """
-
-        self.m_board.sendrshellcmd(f"cp {src} {dst}", timeout = self.m_rshell_fileop_timeout)
-
-#
 # Circuit Python board
 #
 class TestBoardCP(TestBoard):
@@ -356,7 +269,7 @@ class TestBoardCP(TestBoard):
         #
         # pylint seems to prefer this.
         #
-        super().__init__(mountpoint, FileOPsCP())
+        super().__init__(mountpoint, fileops.FileOPsCP())
 
 
     def create_pexepect_child(self):
@@ -404,7 +317,7 @@ class TestBoardMP(TestBoard):
         #
         # pylint seems to prefer this.
         #
-        super().__init__("/pyboard", FileOPsMP(self))
+        super().__init__("/pyboard", fileops.FileOPsMP(self))
 
 
     def create_pexepect_child(self):
